Actually kill processes properly
This switches from `docker-compose pause` to `docker-compose kill` to
kill processes in containers. The container's filesystem is preserved
provided the container is not removed (which must be done explicitly).
This cuases the threshold percentage tests to fail like we expect based
on designate's current behavior (never recovering zones which are in an
ACTIVE status, even if the zone is missing from some nameserver)

(The problem was that `docker-compose pause` only suspends a process.
The kernel (or the network stack) maintains tcp connections for suspended
processes (receiving/acking packets). `docker-compose unpause` resumes
the process, after which the process receives the packets which were
queued during suspension. This falsely gave the appearence of
resiliency in designate, since a nameserver would come back on like and
magically receive all updates.)